requirements.txt
blinker==1.8.2
boto3==1.35.28
botocore==1.35.28
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==2.1.5
numpy==2.1.1
pandas==2.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.0.4